Version 1.5  Recyclers AUTOUPDATED OK

+++ app/src/main/java/com/muvbit/elnoticiero/fragments/FavoriteNewsFragment.kt
@@ -5,9 +5,11 @@ import android.util.Log
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
+import androidx.appcompat.app.AlertDialog
 import androidx.fragment.app.Fragment
 import androidx.lifecycle.lifecycleScope
 import androidx.recyclerview.widget.LinearLayoutManager
+import com.google.android.material.snackbar.Snackbar
 import com.muvbit.elnoticiero.R
 import com.muvbit.elnoticiero.activities.MainActivity
 import com.muvbit.elnoticiero.adapters.FavoriteNewsAdapter
@@ -17,11 +19,9 @@ import com.muvbit.elnoticiero.databinding.FragmentFavoriteNewsBinding
 import com.muvbit.elnoticiero.model.News
 import kotlinx.coroutines.launch
 
-
 class FavoriteNewsFragment : Fragment() {
 
     private lateinit var binding: FragmentFavoriteNewsBinding
-
     private lateinit var newsAdapter: FavoriteNewsAdapter
     private lateinit var favoriteNewsRepository: NewsRepository
     private var favoriteNewsList: MutableList<News> = mutableListOf()
@@ -37,66 +37,66 @@ class FavoriteNewsFragment : Fragment() {
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
 
-
-        //Obtener la actividad principal
+        // Get the main activity
         val mainActivity = requireActivity() as MainActivity
         val mainActivityBinding = mainActivity.binding
 
         mainActivityBinding.bottomNav.menu.clear()
         mainActivityBinding.bottomNav.inflateMenu(R.menu.bottom_favorite_menu)
 
-        //BASE DE DATOS
+        // Database
         val appDatabase = NewsDatabase.getDatabase(requireContext())
         favoriteNewsRepository = NewsRepository(appDatabase.NewsDao())
 
         binding.favoriteNewsRecyclerView.layoutManager = LinearLayoutManager(context)
-        /*
-        lifecycleScope.launch {
-            favoriteNewsRepository.getAllFavoriteNews().collect { favoriteNewsList ->
-                val newsList = favoriteNewsList.map { mapFavoriteNewsToNews(it) }
-                setupRecyclerView(newsList)
-            }
-        }
+        newsAdapter = FavoriteNewsAdapter(favoriteNewsList)
+        binding.favoriteNewsRecyclerView.adapter = newsAdapter
 
-         */
         loadFavoriteNews()
 
         mainActivityBinding.bottomNav.setOnItemSelectedListener { item ->
             when (item.itemId) {
                 R.id.favoriteDeleteAll -> {
-                    lifecycleScope.launch {
-                    favoriteNewsRepository.deleteAllFavoriteNews()
-                        favoriteNewsList.clear()
-                        loadFavoriteNews()
-                    }
+                    AlertDialog.Builder(requireContext())
+                        .setMessage(getString(R.string.areYouSureDeleteAllFavorites))
+                        .setPositiveButton(R.string.yes) { _, _ ->
+                            lifecycleScope.launch {
+                                favoriteNewsRepository.deleteAllFavoriteNews()
+                                loadFavoriteNews()
+                                Snackbar.make(
+                                    binding.root,
+                                    getString(R.string.all_news_deleted_from_favorites),
+                                    Snackbar.LENGTH_SHORT
+                                ).show()
+                            }
+                        }
+                        .setNegativeButton(R.string.no, null)
+                        .show()
                     true
                 }
                 else -> false
             }
         }
-
     }
+
     private fun loadFavoriteNews() {
         lifecycleScope.launch {
-            favoriteNewsList = favoriteNewsRepository.getAllFavoriteNews().toMutableList()
-            setupRecyclerView(favoriteNewsList)
-            }
+            val newFavoriteNewsList = favoriteNewsRepository.getAllFavoriteNews().toMutableList()
+            setupRecyclerView(newFavoriteNewsList)
         }
+    }
 
-
-    private fun setupRecyclerView(newsList: List<News>) {
-        Log.d("FavoriteNewsFragment", "setupRecyclerView called with list size: ${newsList.size}")
-        if (newsList.isNotEmpty()) {
-            newsAdapter = FavoriteNewsAdapter(newsList)
-            binding.favoriteNewsRecyclerView.apply {
-                adapter = newsAdapter
-            }
+    private fun setupRecyclerView(newFavoriteNewsList: MutableList<News>) {
+        Log.d("FavoriteNewsFragment", "setupRecyclerView called with list size: ${newFavoriteNewsList.size}")
+        favoriteNewsList.clear()
+        favoriteNewsList.addAll(newFavoriteNewsList)
+        newsAdapter.updateNewsList(favoriteNewsList)
+        if (favoriteNewsList.isNotEmpty()) {
             binding.favoriteNewsCenterMessage.visibility = View.GONE
         } else {
             binding.favoriteNewsCenterMessage.visibility = View.VISIBLE
         }
     }
-
 
     override fun onDestroyView() {
         super.onDestroyView()
